backend/requirements.txt
Django==4.2.*
djangorestframework==3.14.*
mssql-django==1.5.*
pyodbc==5.1.*
djangorestframework-simplejwt==5.3.*
python-decouple==3.8
drf-spectacular==0.26.*
gunicorn==21.2.*
